Move tempgitdir context manager to main test module
The context manager will be used for another test case.
